# shop-api/src/main/java/com/shop/ws/dto/Status.java
package com.shop.ws.dto;

public class Status {
	private boolean success;
	private String message;

	public Status(boolean success) {
		this.success = success;
	}

	public Status(boolean success, String message) {
		this.success = success;
		this.message = message;
	}

	public boolean isSuccess() {
		return success;
	}

	public void setSuccess(boolean success) {
		this.success = success;
	}

	public String getMessage() {
		return message;
	}

	public void setMessage(String message) {
		this.message = message;
	}

}